+++ 4-governance/dubbo-samples-spring-boot-tracing/dubbo-samples-spring-boot-tracing-consumer/src/main/java/org/apache/dubbo/springboot/demo/consumer/ConsumerApplication.java
