+++ 4-governance/dubbo-samples-spring-boot-tracing/dubbo-samples-spring-boot-tracing-consumer/src/main/java/org/apache/dubbo/springboot/demo/consumer/ObservationConfiguration.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.springboot.demo.provider;
+package org.apache.dubbo.springboot.demo.consumer;
 
 import io.opentelemetry.exporter.zipkin.ZipkinSpanExporterBuilder;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
@@ -29,4 +29,5 @@ public class ObservationConfiguration {
     SpanExporter spanExporter() {
         return new ZipkinSpanExporterBuilder().setEndpoint("http://localhost:9411/api/v2/spans").build();
     }
+
 }
