+++ 4-governance/dubbo-samples-spring-boot-tracing/dubbo-samples-spring-boot-tracing-provider/pom.xml
@@ -19,33 +19,36 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>org.apache.dubbo</groupId>
-        <artifactId>dubbo-samples-spring-boot3-tracing</artifactId>
+        <artifactId>dubbo-samples-spring-boot-tracing</artifactId>
         <version>1.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>dubbo-samples-spring-boot3-tracing-consumer</artifactId>
+
+    <artifactId>dubbo-samples-spring-boot-tracing-provider</artifactId>
 
     <properties>
+        <slf4j-log4j12.version>1.7.25</slf4j-log4j12.version>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-samples-spring-boot3-tracing-interface</artifactId>
+            <artifactId>dubbo-samples-spring-boot-tracing-interface</artifactId>
             <version>${project.parent.version}</version>
-        </dependency>
-
-        <!-- registry dependency -->
-        <dependency>
-            <groupId>com.alibaba.nacos</groupId>
-            <artifactId>nacos-client</artifactId>
         </dependency>
 
         <!-- dubbo starter -->
         <dependency>
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-spring-boot-starter</artifactId>
+        </dependency>
+
+        <!-- registry dependency -->
+        <dependency>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo-dependencies-zookeeper</artifactId>
+            <type>pom</type>
         </dependency>
 
         <!-- spring starter -->
@@ -74,6 +77,7 @@
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
+                <version>${spring-boot-maven-plugin.version}</version>
                 <executions>
                     <execution>
                         <goals>
@@ -84,4 +88,5 @@
             </plugin>
         </plugins>
     </build>
+
 </project>
