+++ 4-governance/dubbo-samples-spring-boot-tracing/dubbo-samples-spring-boot-tracing-provider/src/main/java/org/apache/dubbo/springboot/demo/provider/ObservationConfiguration.java
@@ -29,4 +29,5 @@ public class ObservationConfiguration {
     SpanExporter spanExporter() {
         return new ZipkinSpanExporterBuilder().setEndpoint("http://localhost:9411/api/v2/spans").build();
     }
+
 }
