+++ 4-governance/dubbo-samples-spring-boot-tracing/dubbo-samples-spring-boot-tracing-provider/src/main/java/org/apache/dubbo/springboot/demo/provider/ProviderApplication.java
