+++ 4-governance/pom.xml
@@ -40,7 +40,7 @@
         <module>dubbo-samples-spring-hystrix</module>
         <module>dubbo-samples-ssl</module>
         <module>dubbo-samples-tagrouter</module>
-        <module>dubbo-samples-tracing</module>
+        <module>dubbo-samples-spring-boot-tracing</module>
         <module>dubbo-samples-zipkin</module>
         <module>dubbo-samples-spring-boot3-tracing</module>
         <module>dubbo-samples-spring-boot-tracing-skywalking</module>
